+++ storage/auth.go
@@ -5,7 +5,6 @@ import (
 	"strconv"
 
 	"github.com/dgrijalva/jwt-go"
-	"github.com/user/2019_1_newTeam2/models"
 )
 
 func (db *Database) Login(username string, password string, secret []byte) (string, string, error) {
@@ -30,27 +29,4 @@ func (db *Database) Login(username string, password string, secret []byte) (stri
 
 	}
 	return "", "", fmt.Errorf("Неверное имя пользователя")
-}
-
-func (db *Database) UserRegistration(username string, email string,
-	password string, langid int, pronounceOn int) (bool, error) {
-
-	check, _ := db.CheckUserByUsername(username)
-	if check {
-		return false, fmt.Errorf("Такой пользователь уже существует")
-	}
-
-	for _, i := range db.UserData {
-		if i.Username == username {
-			return false, fmt.Errorf("Такой пользователь уже существует")
-		}
-	}
-	id := db.LastUserId
-	db.Logger.Log(db.LastUserId)
-	_, err := HashPassword(password)
-	if err != nil {
-		return false, fmt.Errorf("hash error")
-	}
-	db.UserData[id] = models.User{id, username, email, password, langid, pronounceOn, 0, "files/avatars/shrek.jpg"}
-	return true, nil
 }
